Changed the USB HID report structure
accomodation for the LUFA Framework.

+++ Milestone Applications/Generic Joystick USB Buttons Demo/CLASS_JOYSTICK1/Joystick.h
@@ -59,11 +59,11 @@
 		 */
 		typedef struct
 		{
-			int8_t  X; /**< Current absolute joystick X position, as a signed 8-bit integer */
-			int8_t  Y; /**< Current absolute joystick Y position, as a signed 8-bit integer */
-			int8_t  Z; /**< Current absolute joystick Z position, as a signed 8-bit integer */
+			int16_t  X; /**< Current absolute joystick X position, as a signed 8-bit integer */
+			int16_t  Y; /**< Current absolute joystick Y position, as a signed 8-bit integer */
+			int16_t  Z; /**< Current absolute joystick Z position, as a signed 8-bit integer */
 			uint8_t Button; /**< Bit mask of the currently pressed joystick buttons */
-		} USB_JoystickReport_Data_t;
+		}USB_JoystickReport_Data_t;
 
 	/* Macros: */
 		/** LED mask for the library LED driver, to indicate that the USB interface is not ready. */
